Remove mock and prod variants
