+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/di/DataModules.kt
@@ -16,16 +16,45 @@
 
 package com.example.android.architecture.blueprints.todoapp.di
 
+import android.content.Context
+import androidx.room.Room
+import com.example.android.architecture.blueprints.todoapp.data.source.DefaultTasksRepository
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksDataSource
+import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
 import com.example.android.architecture.blueprints.todoapp.data.source.local.TasksLocalDataSource
 import com.example.android.architecture.blueprints.todoapp.data.source.local.ToDoDatabase
 import com.example.android.architecture.blueprints.todoapp.data.source.remote.TasksRemoteDataSource
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineDispatcher
+import javax.inject.Qualifier
 import javax.inject.Singleton
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+annotation class RemoteTasksDataSource
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+annotation class LocalTasksDataSource
+
+@Module
+@InstallIn(SingletonComponent::class)
+object RepositoryModule {
+
+    @Singleton
+    @Provides
+    fun provideTasksRepository(
+        @RemoteTasksDataSource remoteDataSource: TasksDataSource,
+        @LocalTasksDataSource localDataSource: TasksDataSource,
+        @IoDispatcher ioDispatcher: CoroutineDispatcher
+    ): TasksRepository {
+        return DefaultTasksRepository(remoteDataSource, localDataSource, ioDispatcher)
+    }
+}
 
 @Module
 @InstallIn(SingletonComponent::class)
@@ -44,5 +73,20 @@ object DataSourceModule {
         @IoDispatcher ioDispatcher: CoroutineDispatcher
     ): TasksDataSource {
         return TasksLocalDataSource(database.taskDao(), ioDispatcher)
+    }
+}
+
+@Module
+@InstallIn(SingletonComponent::class)
+object DatabaseModule {
+
+    @Singleton
+    @Provides
+    fun provideDataBase(@ApplicationContext context: Context): ToDoDatabase {
+        return Room.databaseBuilder(
+            context.applicationContext,
+            ToDoDatabase::class.java,
+            "Tasks.db"
+        ).build()
     }
 }
