--- app/src/main/java/com/example/android/architecture/blueprints/todoapp/di/RepositoryModule.kt
@@ -1,51 +1,0 @@
-/*
- * Copyright (C) 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.android.architecture.blueprints.todoapp.di
-
-import com.example.android.architecture.blueprints.todoapp.data.source.DefaultTasksRepository
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksDataSource
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineDispatcher
-import javax.inject.Qualifier
-import javax.inject.Singleton
-
-@Qualifier
-@Retention(AnnotationRetention.RUNTIME)
-annotation class RemoteTasksDataSource
-
-@Qualifier
-@Retention(AnnotationRetention.RUNTIME)
-annotation class LocalTasksDataSource
-
-@Module
-@InstallIn(SingletonComponent::class)
-object RepositoryModule {
-
-    @Singleton
-    @Provides
-    fun provideTasksRepository(
-        @RemoteTasksDataSource remoteDataSource: TasksDataSource,
-        @LocalTasksDataSource localDataSource: TasksDataSource,
-        @IoDispatcher ioDispatcher: CoroutineDispatcher
-    ): TasksRepository {
-        return DefaultTasksRepository(remoteDataSource, localDataSource, ioDispatcher)
-    }
-}
