--- app/src/mock/java/com/example/android/architecture/blueprints/todoapp/data/FakeTasksRemoteDataSource.kt
@@ -1,109 +1,0 @@
-/*
- * Copyright (C) 2019 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.example.android.architecture.blueprints.todoapp.data
-
-import com.example.android.architecture.blueprints.todoapp.data.Result.Error
-import com.example.android.architecture.blueprints.todoapp.data.Result.Success
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksDataSource
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.map
-
-/**
- * Implementation of a remote data source with static access to the data for easy testing.
- */
-object FakeTasksRemoteDataSource : TasksDataSource {
-
-    private var TASKS_SERVICE_DATA: LinkedHashMap<String, Task> = LinkedHashMap()
-
-    private val observableTasks = MutableStateFlow<Result<List<Task>>>(
-        Success(TASKS_SERVICE_DATA.values.toList())
-    )
-
-    override suspend fun refreshTasks() {
-        observableTasks.value = getTasks()
-    }
-
-    override suspend fun refreshTask(taskId: String) {
-        refreshTasks()
-    }
-
-    override fun getTasksStream(): Flow<Result<List<Task>>> {
-        return observableTasks
-    }
-
-    override fun getTaskStream(taskId: String): Flow<Result<Task>> {
-        return observableTasks.map { tasks ->
-            when (tasks) {
-                is Error -> Error(tasks.exception)
-                is Success -> {
-                    val task = tasks.data.firstOrNull() { it.id == taskId }
-                        ?: return@map Error(Exception("Not found"))
-                    Success(task)
-                }
-            }
-        }
-    }
-
-    override suspend fun getTask(taskId: String): Result<Task> {
-        TASKS_SERVICE_DATA[taskId]?.let {
-            return Success(it)
-        }
-        return Error(Exception("Could not find task"))
-    }
-
-    override suspend fun getTasks(): Result<List<Task>> {
-        return Success(TASKS_SERVICE_DATA.values.toList())
-    }
-
-    override suspend fun saveTask(task: Task) {
-        TASKS_SERVICE_DATA[task.id] = task
-    }
-
-    override suspend fun completeTask(task: Task) {
-        val completedTask = Task(task.title, task.description, true, task.id)
-        TASKS_SERVICE_DATA[task.id] = completedTask
-    }
-
-    override suspend fun completeTask(taskId: String) {
-        // Not required for the remote data source.
-    }
-
-    override suspend fun activateTask(task: Task) {
-        val activeTask = Task(task.title, task.description, false, task.id)
-        TASKS_SERVICE_DATA[task.id] = activeTask
-    }
-
-    override suspend fun activateTask(taskId: String) {
-        // Not required for the remote data source.
-    }
-
-    override suspend fun clearCompletedTasks() {
-        TASKS_SERVICE_DATA = TASKS_SERVICE_DATA.filterValues {
-            !it.isCompleted
-        } as LinkedHashMap<String, Task>
-    }
-
-    override suspend fun deleteTask(taskId: String) {
-        TASKS_SERVICE_DATA.remove(taskId)
-        refreshTasks()
-    }
-
-    override suspend fun deleteAllTasks() {
-        TASKS_SERVICE_DATA.clear()
-        refreshTasks()
-    }
-}
